Trying to fix grappling hook
I got rid of the rigidbody on the player, because I realized that it wouldn't work with the character controller. Instead, I have a function in the updage method in the player controller to simulates velocity over time and pushes the player forward. It isn't working exactly how I want it to, partly because of me trying to simulate a rigidbody, so what I am going to do is rewrite a lot of the grappling hook code so that it starts a grappling hook, plays the hook forward, then stops