I'm still working on the grappling hook
None of the YT videos are going to help me, especially since they use a RB. Instead, I am using chat GPT to help me write the code to make a grappling hook using a character controller. the frusterating aspect of this is that i have to simulate a lot of the physics in order to work, such as the velocity. what is currently wrong is the grapple itself